move `@utility` registration to `./utilities.ts`
